Moving pieces of logic around
