+++ packages/sw-cli/src/lib/constants.js
@@ -1,0 +1,5 @@
+module.exports = {
+  blacklistDirectoryNames: [
+    'node_modules',
+  ]
+};
