+++ packages/sw-cli/src/lib/errors.js
